+++ ogion/backup_targets/folder.py
@@ -4,9 +4,9 @@
 import logging
 from pathlib import Path
 
-from backuper import core
-from backuper.backup_targets.base_target import BaseBackupTarget
-from backuper.models.backup_target_models import DirectoryTargetModel
+from ogion import core
+from ogion.backup_targets.base_target import BaseBackupTarget
+from ogion.models.backup_target_models import DirectoryTargetModel
 
 log = logging.getLogger(__name__)
