+++ ogion/backup_targets/targets_mapping.py
@@ -1,7 +1,7 @@
 # Copyright: (c) 2024, Rafał Safin <rafal.safin@rafsaf.pl>
 # GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
-from backuper.backup_targets import (
+from ogion.backup_targets import (
     base_target,
     file,
     folder,
@@ -9,7 +9,7 @@ from backuper.backup_targets import (
     mysql,
     postgresql,
 )
-from backuper.config import BackupTargetEnum
+from ogion.config import BackupTargetEnum
 
 
 def get_target_cls_map() -> dict[str, type[base_target.BaseBackupTarget]]:
