+++ ogion/bin/7zip/amd64/License.txt
